Adds Android on Chrome OS Google Drive sample showing backup of data using the Google Drive REST API
This sample showcases how to use the Google Drive REST API to backup user data
to give EDU developers on Chrome OS Devices an easy way to solve the problem
of data being deleted due to lack of storage available. This uses Work Manager
to showcase how to update things in the background as well.

Change-Id: Ic67e92e961a4c3d3b57de8d5d5b9759056ae9fd3